Move file to package tscec
